refactor(ci): replace resources.toml/ResourceSync bootstrap with komodo-deploy actions
The ResourceSync approach required komodo/resources.toml to already be
merged to main before Komodo could pull and apply it, which is exactly
why the deploy job was failing on this branch. Ensuring Build/Stack/
Action resources now happens directly from CI via the reusable
BrockCSC/komodo-deploy composite actions, so it self-heals on any
branch without manual setup in the Komodo UI.

Website-specific logic (branch->environment naming, subdomain scheme,
DB schema convention) moves to komodo/deploy-context.mjs; everything
reusable moves to BrockCSC/komodo-deploy.
